using long instead of int

+++ src/main/java/hr/bp/adventofcode/day09/PolynomialEquation.java
@@ -29,13 +29,13 @@ public class PolynomialEquation {
         calculateCoefficients();
     }
 
-    public Integer calculateNextValue() {
-        int result = 0;
+    public Long calculateNextValue() {
+        long result = 0;
 
-        int x = sequenceOfNumbers.size() + 1;
+        long x = sequenceOfNumbers.size() + 1;
 
         for (int i = 0; i < coefficients.getDimension(); i++) {
-            result += (int) Math.round(coefficients.getEntry(i) * Math.pow(x, i));
+            result += Math.round(coefficients.getEntry(i) * Math.pow(x, i));
         }
 
         return result;
